--- client/src/pages/Account/Account.module.css
@@ -1,74 +1,0 @@
-.accountContainer {
-  display: flex;
-  justify-content: center;
-}
-
-.generalContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1rem;
-  margin: 1em;
-  margin-top: 2em;
-  background-color: var(--light);
-  border-radius: 15px;
-  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 6px;
-
-  @media screen and (min-width: 750px) {
-    min-width: 40rem;
-    padding: 1rem 5rem;
-  }
-}
-
-.generalContainer input {
-  border-radius: 5px;
-  padding: 0.5em 1em;
-}
-
-.imgContainer {
-  display: flex;
-  justify-content: center;
-}
-
-.imgContainer img {
-  justify-content: center;
-  width: 9vh;
-  height: 9vh;
-}
-
-.idContainer {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  margin: 0 4rem;
-  gap: 1rem;
-}
-
-.idContainer button {
-  color: var(--light);
-  background-color: var(--secondary-color);
-  border-radius: 10px;
-  padding: 0.3rem 1rem;
-}
-
-.mailContainer {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  margin: 0 3rem;
-  gap: 1rem;
-  padding: 0.3rem 1rem;
-}
-
-.mailContainer img {
-  width: 2em;
-  height: 1.5em;
-}
-
-.fieldContainer {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  margin: 0 4rem;
-  gap: 1rem;
-}
